requirements.txt
amqp==5.1.1
asgiref==3.6.0
bcrypt==4.0.1
black==23.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==36.0.2
csp==0.1a10
defusedxml==0.6.0
distlib==0.3.6
Django==4.2.2
django-allauth==0.43.0
django-ckeditor==6.0.0
django-cleanup==5.1.0
django-cors-headers==3.6.0
django-countries==7.0
django-elasticsearch-dsl==7.1.1
django-js-asset==2.0.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
elasticsearch==7.17.1
elasticsearch-dsl==7.1.0
idna==3.4
jmespath==0.10.0
jwt==1.3.1
kombu==5.2.4
loremipsum==1.0.5
MarkupSafe==2.1.1
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.1
pathspec==0.11.1
Pillow==9.0.1
platformdirs==2.5.4
prompt-toolkit==3.0.28
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.15
uWSGI==2.0.21
vine==5.0.0
wcwidth==0.2.6